+++ BlazorAppTest01/BlazorAppTest01/Program.cs
@@ -4,14 +4,18 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authentication.OpenIdConnect;
 using Microsoft.IdentityModel.Tokens;
 using Rebel.Alliance.Canary.Services;
-using Rebel.Alliance.Canary.Actors;
-using Rebel.Alliance.Canary.Messaging;
 using Rebel.Alliance.Canary.Abstractions;
 using System.Text;
 using MediatR;
 using Rebel.Alliance.Canary.Configuration;
-using Rebel.Alliance.Canary.SystemProviders;
 using Microsoft.Extensions.DependencyInjection;
+using Rebel.Alliance.Canary.InMemoryActorFramework.Actors.CredentialVerifierActor;
+using Rebel.Alliance.Canary.InMemoryActorFramework.Actors.RevocationManagerActor;
+using Rebel.Alliance.Canary.InMemoryActorFramework.Actors.TokenIssuerActor;
+using Rebel.Alliance.Canary.InMemoryActorFramework.SystemProviders;
+using Rebel.Alliance.Canary.InMemoryActorFramework;
+using Rebel.Alliance.Canary.InMemoryActorFramework.Actors.OIDCClientActor;
+using Rebel.Alliance.Canary.Abstractions.Actors;
 
 var builder = WebApplication.CreateBuilder(args);
