+++ Rebel.Alliance.Canary/InMemoryActorFramework/Actors/TrustFrameworkManagerActor/TrustFrameworkManagerActor.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using MediatR;
 using Rebel.Alliance.Canary.Abstractions;
+using Rebel.Alliance.Canary.Abstractions.Actors;
 
-namespace Rebel.Alliance.Canary.Actors
+namespace Rebel.Alliance.Canary.InMemoryActorFramework.Actors.TrustFrameworkManagerActor
 {
     public class TrustFrameworkManagerActor : ActorBase, ITrustFrameworkManagerActor
     {
@@ -82,13 +83,5 @@ namespace Rebel.Alliance.Canary.Actors
             Console.WriteLine($"Issuer is not trusted: {issuerDid}");
             return false;
         }
-    }
-
-    public interface ITrustFrameworkManagerActor : IActor
-    {
-        Task<bool> RegisterIssuerAsync(string issuerDid, string publicKey);
-        Task<bool> CertifyIssuerAsync(string issuerDid);
-        Task<bool> RevokeIssuerAsync(string issuerDid);
-        Task<bool> IsTrustedIssuerAsync(string issuerDid);
     }
 }
