+++ Rebel.Alliance.Canary/InMemoryActorFramework/SystemProviders/InMemoryActorStateManager.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Rebel.Alliance.Canary.Abstractions;
 
-namespace Rebel.Alliance.Canary.SystemProviders
+namespace Rebel.Alliance.Canary.InMemoryActorFramework.SystemProviders
 {
     public class InMemoryActorStateManager : IActorStateManager
     {
