+++ Rebel.Alliance.Canary/InMemoryActorFramework/SystemProviders/InMemorySystemProvider.cs
@@ -1,4 +1,4 @@
-﻿namespace Rebel.Alliance.Canary.SystemProviders
+﻿namespace Rebel.Alliance.Canary.InMemoryActorFramework.SystemProviders
 {
     using System;
     using System.Collections.Concurrent;
