--- Rebel.Alliance.Canary/Messaging/Handlers.cs/ActorMessageHandler.cs
@@ -1,22 +1,0 @@
-﻿using MediatR;
-using Rebel.Alliance.Canary.Messaging;
-using Rebel.Alliance.Canary.Abstractions;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
-
-public class ActorMessageHandler<TActor> : IRequestHandler<ActorMessageEnvelope<TActor>, object> where TActor : IActor
-{
-    private readonly IServiceProvider _serviceProvider;
-
-    public ActorMessageHandler(IServiceProvider serviceProvider)
-    {
-        _serviceProvider = serviceProvider;
-    }
-
-    public async Task<object> Handle(ActorMessageEnvelope<TActor> request, CancellationToken cancellationToken)
-    {
-        var actor = _serviceProvider.GetRequiredService<TActor>();
-        return await actor.ReceiveAsync((IActorMessage)request.Message);
-    }
-}
