+++ Rebel.Alliance.Canary/Messaging/RevokeCredentialMessage.cs
@@ -4,6 +4,13 @@ namespace Rebel.Alliance.Canary.Messaging
 {
     public class RevokeCredentialMessage : IActorMessage
     {
+        public string CredentialId { get; }
+
+        public RevokeCredentialMessage(string credentialId)
+        {
+            CredentialId = credentialId;
+        }
+
         public string MessageType => nameof(RevokeCredentialMessage);
     }
 }
