+++ Rebel.Alliance.Canary/Messaging/ValidateTokenMessage.cs
@@ -1,14 +1,13 @@
 using Rebel.Alliance.Canary.Abstractions;
 public class ValidateTokenMessage : IActorMessage
 {
-    public string ClientId { get; }
     public string Token { get; }
-    public string MessageType { get; }
 
-    public ValidateTokenMessage(string clientId, string token)
+    public ValidateTokenMessage(string token)
     {
-        ClientId = clientId;
         Token = token;
-        MessageType = "ValidateToken";
     }
+
+    public string MessageType => nameof(ValidateTokenMessage);
 }
+
