--- Rebel.Alliance.Canary/Messaging/ValidateTokenMessage1.cs
@@ -1,9 +1,0 @@
-﻿using Rebel.Alliance.Canary.Abstractions;
-
-namespace Rebel.Alliance.Canary.Messaging
-{
-    public class ValidateTokenMessage : IActorMessage
-    {
-        public string MessageType => nameof(ValidateTokenMessage);
-    }
-}
