+++ Rebel.Alliance.Canary/Models/Proof.cs
@@ -1,12 +1,39 @@
 ﻿namespace Rebel.Alliance.Canary.Models
 {
-    public class Proof
+    using System.ComponentModel.DataAnnotations;
+
+    namespace Rebel.Alliance.Canary.Models
     {
-        public string Type { get; set; } = "Ed25519Signature2018";  // The type of cryptographic signature (e.g., EdDSA, RSA, etc.)
-        public DateTime Created { get; set; }  // The date and time when the proof was created
-        public string VerificationMethod { get; set; }  // Reference to the verification method or public key
-        public string ProofPurpose { get; set; } = "assertionMethod";  // The intended purpose of the proof (e.g., authentication, assertion, etc.)
-        public string Jws { get; set; }  // The actual signature in JWS (JSON Web Signature) format
+        public class Proof
+        {
+            [Required]
+            public string Type { get; set; } = "Ed25519Signature2018";
+
+            [Required]
+            public DateTime Created { get; set; }
+
+            [Required]
+            public string VerificationMethod { get; set; }
+
+            [Required]
+            public string ProofPurpose { get; set; } = "assertionMethod";
+
+            [Required]
+            public string Jws { get; set; }
+
+            public string Creator { get; set; }
+
+            public string Domain { get; set; }
+
+            public bool IsValid()
+            {
+                return !string.IsNullOrEmpty(Type) &&
+                       Created != default &&
+                       !string.IsNullOrEmpty(VerificationMethod) &&
+                       !string.IsNullOrEmpty(ProofPurpose) &&
+                       !string.IsNullOrEmpty(Jws);
+            }
+        }
     }
 
 }
